refactor(layout): move book & movie details to ecommerce
