--- src/components/dashboard/index.ts
@@ -1,16 +1,0 @@
-export {
-  BookHeader,
-  BookHeaderProps,
-} from './bookDetails/bookHeader.component';
-export {
-  MovieHeader,
-  MovieHeaderProps,
-} from './movie/movieHeader.component';
-export {
-  MovieDetailList,
-  MovieDetailListProps,
-} from './movie/movieDetailList.component';
-export {
-  MovieScreenshotList,
-  MovieScreenshotListProps,
-} from './movie/movieScreenshotList.component';
