+++ src/components/ecommerce/movieDetails/movieCategoryListItem.component.tsx
