+++ src/components/ecommerce/movieDetails/movieDetailList.component.tsx
