+++ src/components/ecommerce/movieDetails/movieDetailListItem.component.tsx
