+++ src/components/ecommerce/movieDetails/movieScreenshotList.component.tsx
