+++ src/containers/layouts/dashboards/dashboards.container.tsx
@@ -20,16 +20,6 @@ export class DashboardsContainer extends React.Component<NavigationScreenProps> 
       description: 'Trainings 2',
       route: 'Trainings 2',
     },
-    {
-      title: 'Layout',
-      description: 'Movie',
-      route: 'Movie',
-    },
-    {
-      title: 'Layout',
-      description: 'Book Details',
-      route: 'Book Details',
-    },
   ];
 
   private onItemSelect = (index: number): void => {
