+++ src/containers/layouts/dashboards/trainings1/hard1.component.tsx
