+++ src/containers/layouts/dashboards/trainings1/trainings1.component.tsx
