+++ src/containers/layouts/dashboards/trainings1/trainings1.container.tsx
