+++ src/containers/layouts/dashboards/trainings2/trainings2.container.tsx
