+++ src/containers/layouts/ecommerce/movieDetails/movieDetails.container.tsx
@@ -2,20 +2,20 @@ import React from 'react';
 import { NavigationScreenProps } from 'react-navigation';
 import { movie1 } from '@src/core/data/movie';
 import { Movie as MovieModel } from '@src/core/model';
-import { Movie } from './movie.component';
+import { Movie } from './movieDetails.component';
 
 interface State {
   movie: MovieModel;
 }
 
-export class MovieContainer extends React.Component<NavigationScreenProps, State> {
+export class MovieDetailsContainer extends React.Component<NavigationScreenProps, State> {
 
   public state: State = {
     movie: movie1,
   };
 
   private onBuyTicketPress = (): void => {
-
+    this.props.navigation.navigate('Payment');
   };
 
   public render(): React.ReactNode {
