+++ src/containers/layouts/index.ts
@@ -18,10 +18,10 @@ export { SignUp3Container } from './auth/signUp3/signUp3.container';
 export { SignUp4Container } from './auth/signUp4/signUp4.container';
 export { ForgotPasswordContainer } from './auth/forgotPassword/forgotPassword.container';
 export { DashboardsContainer } from './dashboards/dashboards.container';
-export { Trainings1Container } from './dashboards/trainings/trainings1/trainings1.container';
-export { Trainings2Container } from './dashboards/trainings/trainings2/trainings2.container';
-export { MovieContainer } from './dashboards/movie/movie.container';
-export { BookDetailsContainer } from './dashboards/bookDetails/bookDetails.container';
+export { Trainings1Container } from './dashboards/trainings1/trainings1.container';
+export { Trainings2Container } from './dashboards/trainings2/trainings2.container';
+export { MovieDetailsContainer } from './ecommerce/movieDetails/movieDetails.container';
+export { BookDetailsContainer } from './ecommerce/bookDetails/bookDetails.container';
 export { EcommerceContainer } from './ecommerce/ecommerce.container';
 export { RentApartmentContainer } from './ecommerce/rentApartment/rentApartment.container';
 export { AddNewCardContainer } from './ecommerce/addNewCard/addNewCard.container';
